pyproject.toml
[project]
name = "mcp-getting-started"
version = "0.1.0"
description = "Getting started with MCP"
authors = []
dependencies = [
    "mcp>=0.4",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]
readme = "README.md"
